refactor: social service to common
