+++ src/app/projects/projects-page/project-list-item/project-list-item.component.ts
@@ -3,8 +3,8 @@ import { ProjectListItem } from '../../project-list-item'
 import { SwiperOptions } from 'swiper/types'
 import { ImageResponsiveBreakpointsService } from '../../../common/images/image-responsive-breakpoints.service'
 import { Author, AuthorsService } from '../../../common/authors.service'
-import { Social } from '../../../about-page/social/social'
-import { SocialService } from '../../../about-page/social/social.service'
+import { Social } from '../../../common/social/social'
+import { SocialService } from '../../../common/social/social.service'
 import { Credit } from '../../credit'
 import { PROJECTS_PATH } from '../../../common/routing/paths'
